user_api/requirements.txt
fastapi
uvicorn[standard]
pymongo
pydantic
python-jose
motor
python-dotenv
python-decouple
motor
pydantic[email]
passlib[bcrypt]